Bevy v0.5 updates

+++ src/window_size.rs
@@ -6,14 +6,16 @@ use bevy::{prelude::*, render::camera::Camera};
 /// Hopefully this may be integrated into bevy in future.
 #[derive(Default)]
 pub struct WindowSize {
-    pub height: u32,
-    pub width: u32,
+    pub height: f32,
+    pub width: f32,
 }
 
 /// Syncs the window width to the camera
 pub fn window_size(windows: Res<Windows>, mut camera_query: Query<(&Camera, &mut WindowSize)>) {
     for (cam, mut size) in camera_query.iter_mut() {
         if let Some(window) = windows.get(cam.window) {
+            // Note: This conversion will be lossy, but u32 is used internally
+            // for texture calcs.
             let windowsize = WindowSize {
                 height: window.height(),
                 width: window.width(),
